+++ Pandora/Paradigm/Structure/Interface/Set.hs
@@ -11,12 +11,12 @@ import Pandora.Paradigm.Primary.Functor.Predicate (equate)
 import Pandora.Paradigm.Primary.Functor.Product (attached)
 import Pandora.Paradigm.Primary.Object.Boolean (Boolean (True, False))
 import Pandora.Paradigm.Primary.Object.Natural (Natural (Zero))
-import Pandora.Paradigm.Structure.Ability.Monotonic (Monotonic (bypass), find)
+import Pandora.Paradigm.Structure.Ability.Monotonic (Monotonic (reduce), find)
 import Pandora.Paradigm.Inventory.State (State, modify)
 import Pandora.Paradigm.Controlflow.Effect (run)
 
 member :: forall e a . (Setoid a, Monotonic e a) => a -> e -> Boolean
-member x = bypass @(Maybe a) @a (\_ _ -> True) False . find (equate x)
+member x = reduce @(Maybe a) @a (\_ _ -> True) False . find (equate x)
 
 subset :: (Monotonic (t a) a, Traversable t, Setoid a, Setoid (t a)) => t a -> t a -> Boolean
 subset ss s = Nothing /= (ss ->> find % s . equate)