+++ CHANGELOG.md
@@ -292,3 +292,4 @@
 * Add `>-<` operator to `Invariant` typeclass
 * Add `:~.` type synonymous for polymorphic lens
 * Replace `maybe` with more generic `bypass`
+* Rename `bypass` method of `Monotonic` to `reduce`